Split components into base class, and styled class
This way, when components are extended, they do not automatically bring
forward styling that may not be useful anymore.
