+++ packages/tab-indicator/src/mwc-tab-indicator.ts
@@ -14,12 +14,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import {BaseElement, html, property, query, customElement, PropertyValues, classMap, addHasRemoveClass} from '@material/mwc-base/base-element.js';
-import MDCTabIndicatorFoundation from '@material/tab-indicator/foundation';
-import MDCSlidingTabIndicatorFoundation from '@material/tab-indicator/sliding-foundation.js';
-import MDCFadingTabIndicatorFoundation from '@material/tab-indicator/fading-foundation.js';
+import {TabIndicatorBase} from './mwc-tab-indicator-base.js';
+import {customElement} from '@material/mwc-base/base-element.js';
 import {style} from './mwc-tab-indicator-css.js';
-import { MDCTabIndicatorAdapter } from '@material/tab-indicator/adapter';
 
 declare global {
   interface HTMLElementTagNameMap {
@@ -28,65 +25,6 @@ declare global {
 }
 
 @customElement('mwc-tab-indicator' as any)
-export class TabIndicator extends BaseElement {
-
-  protected mdcFoundation!: MDCTabIndicatorFoundation;
-
-  protected get mdcFoundationClass() {
-    return this.fade ? MDCFadingTabIndicatorFoundation : MDCSlidingTabIndicatorFoundation;
-  }
-
-  @query('.mdc-tab-indicator')
-  protected mdcRoot!: HTMLElement;
-
-  @query('.mdc-tab-indicator__content')
-  protected contentElement!: HTMLElement;
-
-  @property()
-  icon = '';
-
-  @property({type: Boolean})
-  fade = false;
-
+export class TabIndicator extends TabIndicatorBase {
   static styles = style;
-
-  render() {
-    const contentClasses = {
-      'mdc-tab-indicator__content--icon': this.icon,
-      'material-icons': this.icon,
-      'mdc-tab-indicator__content--underline': !this.icon
-    };
-    return html`
-      <span class="mdc-tab-indicator ${classMap({'mdc-tab-indicator--fade': this.fade})}">
-        <span class="mdc-tab-indicator__content ${classMap(contentClasses)}">${this.icon}</span>
-      </span>
-      `;
-  }
-
-  updated(changedProperties: PropertyValues) {
-    if (changedProperties.has('fade')) {
-      this.createFoundation();
-    }
-  }
-
-  createAdapter(): MDCTabIndicatorAdapter {
-    return {
-      ...addHasRemoveClass(this.mdcRoot),
-      computeContentClientRect: () => this.contentElement.getBoundingClientRect(),
-      setContentStyleProperty: (prop: string, value: string) =>
-          this.contentElement.style.setProperty(prop, value)
-    };
-  }
-
-  computeContentClientRect() {
-    return this.mdcFoundation.computeContentClientRect();
-  }
-
-  activate(previousIndicatorClientRect?: ClientRect) {
-    this.mdcFoundation.activate(previousIndicatorClientRect);
-  }
-
-  deactivate() {
-    this.mdcFoundation.deactivate();
-  }
 }
